rename buiilder to artifact
